examples/uv-setup/pyproject.toml
[project]
name = "uv-setup-example"
version = "0.1.0"
description = "Minimal UV setup example with FastAPI and Air"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "air",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"